feat: add kustomize (#35)
This pull request adds support for the `kustomize` binary to the
project. The changes include updating configuration files,
documentation, and code to integrate `kustomize` as a new binary option.

### Integration of `kustomize` binary:

* **Configuration Updates:**
  - Added `kustomize` to the list of binaries in `.bin/b.yaml`.

* **Documentation Updates:**
- Updated `README.md` to include a description and link for `kustomize`
under the list of prepackaged binaries.

* **Codebase Updates:**
- Added `kustomize` import in `cmd/b/main.go` and updated the `main`
function to include `kustomize.Binary` in the list of binaries.
[[1]](diffhunk://#diff-9220de3f3af10d1a2eb38298eefa2c9cc92852596770de03bc85dbbb6d136f9fR19)
[[2]](diffhunk://#diff-9220de3f3af10d1a2eb38298eefa2c9cc92852596770de03bc85dbbb6d136f9fR52)
- Created a new file, `pkg/binaries/kustomize/kustomize.go`, which
defines the `Binary` function for `kustomize`. This function specifies
how to fetch, configure, and execute the `kustomize` binary, including
handling its version and URL.