feat: add sops (#40)
This pull request introduces support for the `sops` binary, including
its integration into the project and relevant documentation updates. The
most important changes are grouped into the following themes:

### Integration of `sops` binary:

* Added a new file `pkg/binaries/sops/sops.go` defining the `Binary`
function for managing the `sops` binary, including its versioning and
GitHub repository information.
* Updated `cmd/b/main.go` to import the `sops` package and include
`sops.Binary` in the list of binaries initialized in the `main`
function.
[[1]](diffhunk://#diff-9220de3f3af10d1a2eb38298eefa2c9cc92852596770de03bc85dbbb6d136f9fR21)
[[2]](diffhunk://#diff-9220de3f3af10d1a2eb38298eefa2c9cc92852596770de03bc85dbbb6d136f9fR56)

### Documentation updates:

* Updated `README.md` to include `sops` in the list of prepackaged
binaries, with a link to its GitHub repository.

### Configuration updates:

* Added `sops` to the list of binaries in `.bin/b.yaml`.