feat: add argocd binary support and fix packer environment setup (#87)
This pull request adds support for managing the `argocd` binary within
the project. The main changes include introducing a new package for
`argocd`, updating the application to recognize and handle this binary,
and documenting its addition. There are also minor adjustments to
environment variable handling for consistency.

**New binary support:**

* Added a new package `pkg/binaries/argocd` with an implementation for
managing the `argocd` binary, including logic for version detection and
GitHub release fetching.
* Registered the new `argocd` binary in the main application by
importing the package and including it in the binaries list in
`cmd/b/main.go`.
[[1]](diffhunk://#diff-9220de3f3af10d1a2eb38298eefa2c9cc92852596770de03bc85dbbb6d136f9fR8)
[[2]](diffhunk://#diff-9220de3f3af10d1a2eb38298eefa2c9cc92852596770de03bc85dbbb6d136f9fR52)
* Updated the VSCode launch configuration to use `argocd` as an argument
for debugging.

**Documentation:**

* Added `argocd` to the list of supported binaries in `README.md` with a
brief description and link to the upstream repository.

**Code consistency:**

* Updated environment variable assignment in the `packer` binary
implementation for consistency with the new `argocd` approach.